Another attempt to fix the printf issue on mac: don't use uint64 :)

+++ src/mc/explo/udpor/UnfoldingEvent.cpp
@@ -59,8 +59,8 @@ std::string UnfoldingEvent::to_string() const
   dependencies_string += "]";
 
   return xbt::string_printf("Event %lu, Actor %ld: %s (%lu dependencies: %s)", this->id, associated_transition->aid_,
-                            associated_transition->to_string().c_str(),
-                            static_cast<long unsigned>(immediate_causes.size()), dependencies_string.c_str());
+                            associated_transition->to_string().c_str(), immediate_causes.size(),
+                            dependencies_string.c_str());
 }
 
 EventSet UnfoldingEvent::get_history() const
